+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Menu to change the `config.js`
     - Modules can be installed, added, removed, configured
     - There will be backups of the five last versions of the `config.js` in the `config` folder
+- NOTIFICATION action, see [README.md](README.md#notification-request) for details
 
 ### Changed
 - Smaller font sizes in lists